fix(auth): 规范 Review 修复 — 6 项严重问题
API 设计规范:
- auth.py: 使用统一异常 AuthUnauthorizedError（code=AUTH_UNAUTHORIZED）
- 错误码遵循 {MODULE}_{ERROR_TYPE} 格式

后端结构规范:
- auth.py: 路由层不再做 JWT 解码，改用 Depends(get_current_user_id)
- 认证逻辑移入 FastAPI 依赖注入体系
- exceptions.py: 新增 AuthUnauthorizedError / AuthForbiddenError

MySQL P3C:
- user.py: status 字段 Integer → SmallInteger（unsigned tinyint）

前端结构规范:
- login-settings.tsx: fetch() → apiClient + AUTH_ENDPOINTS 集中管理
- api-client.ts: 请求拦截器自动注入 Authorization: Bearer token
- endpoints.ts: 新增 AUTH_ENDPOINTS 端点常量

+++ backend/app/models/user.py
@@ -1,5 +1,5 @@
 """用户模型"""
-from sqlalchemy import Column, String, Integer, BigInteger
+from sqlalchemy import Column, String, SmallInteger
 from app.models.base import BaseModel
 
 
@@ -12,4 +12,4 @@ class User(BaseModel):
     nickname = Column(String(64), nullable=False, default="", comment="昵称")
     avatar = Column(String(512), nullable=False, default="", comment="头像 URL")
     role = Column(String(32), nullable=False, default="user", comment="角色: admin / user")
-    status = Column(Integer, nullable=False, default=1, comment="状态: 1=启用 0=禁用")
+    status = Column(SmallInteger, nullable=False, default=1, comment="状态: 1=启用 0=禁用")
